+++ pnp/plugins/pull/mqtt.py
@@ -4,7 +4,7 @@ from . import PullBase
 from ...utils import auto_str_ignore
 
 
-@auto_str_ignore(['password'])
+@auto_str_ignore(['_client', 'password'])
 class Subscribe(PullBase):
     """
     Pulls messages from the specified topic from the given mosquitto mqtt