+++ CHANGELOG.md
@@ -6,8 +6,9 @@ You are encouraged to specify explicitly the version in your dependency tools, e
 
     pip install pnp==0.10.0
 
-**Next release (probably 0.18.1)**
+**Next release (~0.18.1)**
 * Breaking: Fixes `ignore_overflow` of `pull.sensor.Sound` plugin (which actually has the opposite effect)
+* Adds slack logging handler to log messages to a slack channel and optionally ping users
 * Adds `pull.net.PortProbe` plugin to probe a specific port if it's being used
 
 **0.18.0**